backend/pyproject.toml
[project]
name = "gamegame"
version = "0.1.0"
description = "LLM-powered board game rules assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "nanoid>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "pgvector>=0.3.6",
    "openai>=1.57.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.17",
    "saq[web]>=0.22.0",
    "redis>=5.2.0",
    "mistralai>=1.2.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "email-validator>=2.2.0",
    "pillow>=11.0.0",
    "aiofiles>=25.1.0",
    "aiosmtplib>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "factory-boy>=3.3.0",
    "ruff>=0.8.0",
    "ty>=0.0.2",
    "pre-commit>=4.0.0",
]

[project.scripts]
gamegame = "gamegame.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gamegame"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "factory-boy>=3.3.0",
    "ruff>=0.8.0",
    "ty>=0.0.2",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # pylint
    "PERF",   # perflint
    "RUF",    # ruff-specific
]
ignore = [
    "E501",     # line too long (handled by formatter)
    "B008",     # function call in default argument (required by FastAPI Query/Depends)
    "PLR0912",  # too many branches (acceptable for complex search/pipeline functions)
    "PLR0915",  # too many statements (acceptable for complex search/pipeline functions)
    "PLR0913",  # too many arguments
    "PLR2004",  # magic value comparison
    "PLC0415",  # import not at top level (intentional for CLI lazy loading)
]

[tool.ruff.lint.per-file-ignores]
"src/gamegame/cli/*" = ["PLC0415"]  # CLI uses lazy imports intentionally
"tests/*" = ["ARG001"]  # Test fixtures need to be in signature even if unused

[tool.ruff.lint.isort]
known-first-party = ["gamegame"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["src", "tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/gamegame"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
